move things around to avoid duplication.
