+++ api/routers/core_dynamic.py
@@ -9,7 +9,7 @@ import schemas
 from chalicelib.core import sessions, errors, errors_viewed, errors_favorite, sessions_assignments, heatmaps, \
     sessions_favorite, assist, sessions_notes, click_maps, sessions_replay, signup, feature_flags
 from chalicelib.core import sessions_viewed
-from chalicelib.core import tenants, users, projects, license, tags
+from chalicelib.core import tenants, users, projects, license
 from chalicelib.core import webhook
 from chalicelib.core.collaboration_slack import Slack
 from chalicelib.utils import captcha, smtp
@@ -551,22 +551,3 @@ def update_feature_flag_status(project_id: int, feature_flag_id: int,
                                data: schemas.FeatureFlagStatus = Body(...)):
     return {"data": feature_flags.update_feature_flag_status(project_id=project_id, feature_flag_id=feature_flag_id,
                                                              is_active=data.is_active)}
-
-# tags
-
-@app.post('/{projectId}/tags', tags=["tags"])
-def tags_create(projectId: int, data: schemas.TagCreate = Body(), context: schemas.CurrentContext = Depends(OR_context)):
-    data = tags.create_tag(project_id=projectId, data=data)
-    return {'data': data}
-
-
-@app.get('/{projectId}/tags', tags=["tags"])
-def tags_list(projectId: int, context: schemas.CurrentContext = Depends(OR_context)):
-    data = tags.list_tags(project_id=projectId)
-    return {'data': data}
-
-
-@app.delete('/{projectId}/tags/{tagId}', tags=["tags"])
-def tags_delete(projectId: int, context: schemas.CurrentContext = Depends(OR_context)):
-    data = tags.delete_tag(tag_id=tagId)
-    return {'data': data}
